requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pillow==11.3.0
jinja2==3.1.6
python-multipart==0.0.20
python-dotenv==1.2.1
requests==2.32.5
pillow-heif==1.1.1
rawpy==0.26.1